Speed up the test suite without losing coverage
- Load the Sentry SDK only when SENTRY_URL is set (mirrors the lazy
  Stripe SDK): every test file re-evaluates the app module graph, so the
  SDK's import-time evaluation was paid hundreds of times per run — and
  it also cuts production cold-boot work.
- Replace @std/expect's toContain via expect.extend: the built-in
  pretty-prints the entire searched value even when the assertion
  passes (~35ms per call on a rendered page); the override formats only
  on failure. Regression tests cover the message paths.
- Drop the in-suite `deno task cpd` subprocess test — the same 0%
  duplication gate already runs as its own step in precommit and CI, so
  the suite re-ran a minute of jscpd for an existing gate.
- Shard the migration restore/chain suites across seven files driven by
  shared factories (index % shardCount), so ~78s of sequential DDL
  spreads across parallel workers instead of serialising on one.
- Add withVirtualBackoff (FakeTime) and use it in the write-lock and
  migration verify retry tests, which genuinely slept out the
  50/150/350ms backoffs.
- Cache the admin guide render once per suite via new cachedAdminPage
  ("render once, assert many"); config-altering tests still fetch their
  own copy.
- Seed pagination filler attendees by cloning one production-created
  booking in a single batch instead of 100 sequential full bookings.
- Pin fast PBKDF2 in the hashing tests and assert the production
  iteration count via getPbkdf2Iterations() instead of really running
  600k iterations.
- Document the fast-test patterns in AGENTS.md.
