+++ test/lib/db/migration-restore/chain-a.test.ts
@@ -1,0 +1,3 @@
+import { defineChainSuite } from "./helpers.ts";
+
+defineChainSuite(0, 4);
