+++ test/lib/db/migration-restore/chain-b.test.ts
@@ -1,0 +1,3 @@
+import { defineChainSuite } from "./helpers.ts";
+
+defineChainSuite(1, 4);
