+++ test/lib/db/migration-restore/chain-c.test.ts
@@ -1,0 +1,3 @@
+import { defineChainSuite } from "./helpers.ts";
+
+defineChainSuite(2, 4);
