+++ test/lib/db/migration-restore/chain-d.test.ts
@@ -1,0 +1,3 @@
+import { defineChainSuite } from "./helpers.ts";
+
+defineChainSuite(3, 4);
