+++ test/lib/db/migration-restore/restore-a.test.ts
@@ -1,0 +1,3 @@
+import { defineRestoreCasesSuite } from "./helpers.ts";
+
+defineRestoreCasesSuite(0, 2);
