+++ test/lib/db/migration-restore/restore-b.test.ts
@@ -1,0 +1,3 @@
+import { defineRestoreCasesSuite } from "./helpers.ts";
+
+defineRestoreCasesSuite(1, 2);
