+++ test/test-utils.ts
@@ -1,3 +1,7 @@
+// Side effect: replace @std/expect's toContain with one that doesn't
+// pretty-print the whole value on success (see fast-expect.ts).
+import "./test-utils/fast-expect.ts";
+
 export * from "./test-utils/activity-log.ts";
 export * from "./test-utils/api-schemas.ts";
 export * from "./test-utils/arrays.ts";
@@ -33,4 +37,5 @@ export * from "./test-utils/settings.ts";
 export { TestBrowser } from "./test-utils/test-browser.ts";
 export * from "./test-utils/test-image.ts";
 export * from "./test-utils/validation.ts";
+export * from "./test-utils/virtual-time.ts";
 export * from "./test-utils/webhooks.ts";
